Refactor index into separate routes #26
