+++ public/settings/edit.html
@@ -4,7 +4,7 @@
             <div class="modal-header">
                 <h5 class="modal-title">Edit game</h5>
             </div>
-            <form action="/api/update" id="editForm" method="post" class="needs-validation" novalidate>
+            <form action="/api/gameEntry/update" id="editForm" method="post" class="needs-validation" novalidate>
                 <input type="hidden" id="editId" name="id">
                 <input type="hidden" id="editIgdbId" name="igdb_id">
                 <div class="modal-body">
